+++ test/controllers/search_controller_test.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+
+# Tests to assure that search results are returned.
+class SearchControllerTest < ActionController::TestCase
+  test 'should get index' do
+    get :index
+    assert_response :success
+  end
+end