+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Added better HTML page titles for blog index and blog posts
   - A count of total blog posts, and blog posts by category is now displayed in
     the blog menu
+  - Started work on providing a search results page for MyApnea and the blog
 - **Gem Changes**
   - Updated to Ruby 2.3.1
   - Updated to devise 4.1.0